ENH Default noisy to False and return replies instead
Instead of printing to the screen and requiring lookup of replies as instance attributes, return the reply directly and default `noisy=False`. The returned data are still available as instance attributes, and they will still be printed to the screen when used in the interpreter if not stored by the user. Returning data from `get_*` methods conforms to standard usage and is more convenient for programmers.